+++ splunk/com/splunk/HttpException.java
@@ -19,9 +19,7 @@ package com.splunk;
 import org.w3c.dom.Document;
 import org.w3c.dom.NodeList;
 
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.io.UnsupportedEncodingException;
+import java.io.*;
 
 /**
  * Thrown for HTTP responses that return an error status code.
@@ -45,43 +43,36 @@ public class HttpException extends RuntimeException {
     static HttpException create(ResponseMessage response) {
         int status = response.getStatus();
 
-        response.getContent().mark(10000);
+        StringBuilder s = new StringBuilder();
+        InputStreamReader r;
+        try {
+            r = new InputStreamReader(response.getContent(), "UTF-8");
+        } catch (UnsupportedEncodingException e1) {
+            throw new AssertionError("How does your system not support UTF-8?");
+        }
+
+        int c = -1;
+        while (true) {
+            try {
+                c = r.read();
+            } catch (IOException e1) {
+                // Not much to be done here if that stream is bad...
+            }
+            if (c == -1) break;
+
+            s.appendCodePoint(c);
+        }
+
         String detail = "";
         try {
             // Attempt to read the error detail from the error response content as XML
-            Document document = Xml.parse(response.getContent());
+            Document document = Xml.parse(new ByteArrayInputStream(detail.getBytes()));
             NodeList msgs = document.getElementsByTagName("msg");
             if (msgs.getLength() > 0)
                 detail = msgs.item(0).getTextContent();
         }
         catch (Exception e) {
             // Not an XML document; return the raw string.
-            try {
-                response.getContent().reset();
-            } catch (IOException e2) {
-                // Not much to be done here if that stream is bad...
-            }
-
-            StringBuilder s = new StringBuilder();
-            InputStreamReader r;
-            try {
-                r = new InputStreamReader(response.getContent(), "UTF-8");
-            } catch (UnsupportedEncodingException e1) {
-                throw new AssertionError("How does your system not support UTF-8?");
-            }
-
-            int c = -1;
-            while (true) {
-                try {
-                    c = r.read();
-                } catch (IOException e1) {
-                    // Not much to be done here if that stream is bad...
-                }
-                if (c == -1) break;
-
-                s.appendCodePoint(c);
-            }
-
             detail = s.toString();
         }