+++ splunk/com/splunk/Index.java
@@ -16,7 +16,6 @@
 
 package com.splunk;
 
-import java.lang.UnsupportedOperationException;
 import java.io.IOException;
 import java.io.OutputStream;
 import java.net.Socket;